fixed reversed antiad options

+++ listeners/antiAdvertisement.js
@@ -23,8 +23,11 @@ module.exports = class antiAdvertisement extends Listener {
         // Delete function, check if msg matches regex
         async function del() {
             if (await message.content.match(regex)) {
+                if (message.guild.me.user.id === message.author.id) return;
                 await message.delete().catch(e => { })
-                if (cachedGuild.warn === 'true') await message.channel.send({ content: `${message.author}, links are not allowed.` }).then(e => e.delete({ timeout: 5000 }).catch(e => { }));
+                if (cachedGuild.warn === 'true') await message.channel.send({ content: `${message.author}, links are not allowed.` }).then(
+                    async e => setTimeout(async () => { await e.delete().catch(e => { }) }, 5000)
+                );
             }
         }
